requirements.txt
torch>=2.3.0
transformers>=4.44.0
accelerate>=0.33.0
huggingface_hub>=0.24.0
pydantic>=2.7.0
PyYAML>=6.0.0
safetensors>=0.4.2


